Add files via upload
In this project, I was tasked with presenting a report to management with a recommendation on how to handle our SOC. I provided a cost-benefit analysis of outsourcing vs maintaining an in house SOC. I also compared using paid software (Splunk) vs open source software (ELK Stack). This report granted me experience in product research, decision making, and clear and concise writing. In the course of my research I also learned about the features available in Splunk and the ELK Stack, as well as how outsourcing various aspects of a SOC is done.